+++ app/src/main/java/com/example/wellnessjournal/ui/home/HomeViewModel.kt
@@ -1,15 +1,10 @@
 package com.example.wellnessjournal.ui.home
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
 class HomeViewModel : ViewModel() {
-    // Commenting this for now since I may learn from it/ use it later
+
    /* private val _text = MutableLiveData<String>().apply {
         value = "This is home Fragment"
     }
     val text: LiveData<String> = _text*/
-
-
 }
