+++ app/src/main/java/com/example/wellnessjournal/ui/welcome/WelcomeViewModel.kt
@@ -1,0 +1,13 @@
+package com.example.wellnessjournal.ui.welcome
+
+import androidx.lifecycle.ViewModel
+
+class WelcomeViewModel : ViewModel() {
+    // Commenting this for now since I may learn from it/ use it later
+   /* private val _text = MutableLiveData<String>().apply {
+        value = "This is welcome Fragment"
+    }
+    val text: LiveData<String> = _text*/
+
+
+}
