+++ app/src/main/res/drawable/home_navigation_icon.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="80dp"
+    android:height="61dp"
+    android:viewportWidth="80"
+    android:viewportHeight="61">
+  <path
+      android:pathData="M32,60.092V39.04H48V60.092H68V32.022H80L40,0.444L0,32.022H12V60.092H32Z"
+      android:fillColor="#495AB6"/>
+</vector>
