+++ app/src/main/res/values/strings.xml
@@ -1,7 +1,7 @@
 <resources>
     <string name="app_name">WellnessJournal</string>
+    <string name="title_welcome">Welcome</string>
     <string name="title_home">Home</string>
-    <string name="title_dashboard">Dashboard</string>
     <string name="title_notifications">Notifications</string>
     <string name="Hello">Hello!</string>
     <string name="welcome_message">Welcome to Fitness Journal, where you can customize fitness
